Add setup.py, restructure files
