+++ __init.py__
@@ -1,0 +1,4 @@
+name = "razer-cli"
+
+from .util import *
+from .settings import *
