+++ razer_cli/settings.py
